requirements.txt
numpy
torch
torchvision
open_clip_torch
salesforce-lavis
ultralytics
